fix: resolve circular import for ApiKeyDialog
