+++ gui_pyside6/ui/__init__.py
@@ -5,6 +5,7 @@ from .settings_dialog import SettingsDialog
 from .tools_panel import ToolsPanel
 from .plugin_manager_dialog import PluginManagerDialog
 from .agent_editor_dialog import AgentEditorDialog
+from .api_key_dialog import ApiKeyDialog
 
 __all__ = [
     "MainWindow",
@@ -12,4 +13,5 @@ __all__ = [
     "ToolsPanel",
     "PluginManagerDialog",
     "AgentEditorDialog",
+    "ApiKeyDialog",
 ]
